Fix debug flag with python 3
