+++ coreapi_cli/main.py
@@ -1,6 +1,7 @@
 from coreapi.compat import b64encode, force_bytes, string_types, urlparse
 from coreapi_cli import __version__ as client_version
 from coreapi_cli import codec_plugins
+from coreapi_cli.debug import DebugSession
 from coreapi_cli.history import History, dump_history, load_history
 import click
 import coreapi
@@ -65,66 +66,19 @@ def get_document_string(doc):
     return '<%s %s>' % (doc.title, json.dumps(doc.url))
 
 
-def debug_request(request):
-    def request_echo(fmt, *args):
-        click.echo(click.style('> ', fg='blue') + (fmt % args))
-
-    headers = request.headers
-    headers['host'] = urlparse.urlparse(request.url).hostname
-
-    request_echo(click.style('%s %s HTTP/1.1', bold=True), request.method, request.path_url)
-    for key, value in sorted(headers.items()):
-        request_echo('%s: %s', key.title(), value)
-    if request.body:
-        request_echo('')
-        for line in request.body.splitlines():
-            request_echo(line)
-
-
-def debug_response(response):
-    def success_echo(fmt, *args):
-        click.echo(click.style('< ', fg='green') + (fmt % args))
-
-    def failure_echo(fmt, *args):
-        click.echo(click.style('< ', fg='red') + (fmt % args))
-
-    def info_echo(fmt, *args):
-        click.echo(click.style('< ', fg='yellow') + (fmt % args))
-
-    response_class = ('%s' % response.status_code)[0] + 'xx'
-    if response_class == '2xx':
-        response_echo = success_echo
-    elif response_class in ('4xx', '5xx'):
-        response_echo = failure_echo
-    else:
-        response_echo = info_echo
-
-    response_echo(click.style('%d %s', bold=True), response.status_code, response.reason)
-    for key, value in sorted(response.headers.items()):
-        response_echo('%s: %s', key.title(), value)
-    if response.content:
-        response_echo('')
-        for line in response.content.splitlines():
-            response_echo(line)
-
-    click.echo()
-
-
 def get_client(decoders=None, debug=False):
     credentials = get_credentials()
     headers = get_headers()
+    session = None
     if debug:
-        callbacks = {
-            'request_callback': debug_request,
-            'response_callback': debug_response
-        }
-    else:
-        callbacks = {}
+        session = DebugSession()
 
     if decoders is None:
         decoders = list(codec_plugins.decoders.values())
 
-    http_transport = coreapi.transports.HTTPTransport(credentials, headers, **callbacks)
+    http_transport = coreapi.transports.HTTPTransport(
+        credentials=credentials, headers=headers, session=session
+    )
     return coreapi.Client(decoders=decoders, transports=[http_transport])
